+++ frontend/src/views/hq/HqOrderManageView.vue
@@ -17,7 +17,7 @@
       </div>
     </div>
 
-    <!-- Tabs (underline style) -->
+    <!-- Tabs -->
     <div class="flex border-b border-gray-200">
       <button v-for="tab in tabs" :key="tab.id"
         @click="setOrderViewTab(tab.id)"
@@ -34,463 +34,35 @@
       </button>
     </div>
 
-    <!-- Tab: 자동 발주 제안 -->
-    <div v-if="activeTab === 'auto'" class="space-y-4">
-      <div class="bg-white border border-gray-200 rounded-lg overflow-hidden">
-        <table class="w-full text-sm text-left">
-          <thead>
-            <tr class="border-b border-gray-200 bg-gray-50">
-              <th class="px-5 py-3 text-[10px] font-bold text-gray-400 uppercase tracking-wider">발주번호</th>
-              <th class="px-5 py-3 text-[10px] font-bold text-gray-400 uppercase tracking-wider">대상 매장</th>
-              <th class="px-5 py-3 text-[10px] font-bold text-gray-400 uppercase tracking-wider">발주일시</th>
-              <th class="px-5 py-3 text-[10px] font-bold text-gray-400 uppercase tracking-wider">총 수량</th>
-              <th class="px-5 py-3 text-[10px] font-bold text-gray-400 uppercase tracking-wider">총 금액</th>
-              <th class="px-5 py-3 text-[10px] font-bold text-gray-400 uppercase tracking-wider">상태</th>
-            </tr>
-          </thead>
-          <tbody class="divide-y divide-gray-100">
-            <tr
-              v-for="o in autoOrders"
-              :key="o.id"
-              class="hover:bg-gray-50/50 transition-colors cursor-pointer"
-              @click="openDetail({ id: o.id, type: '자동', store: o.store, status: o.status, date: o.date, items: o.items })"
-            >
-              <td class="px-5 py-3.5 font-mono text-xs text-gray-400">{{ o.id }}</td>
-              <td class="px-5 py-3.5 font-semibold text-gray-900">{{ o.store }}</td>
-              <td class="px-5 py-3.5 text-xs text-gray-400 font-mono">{{ o.date ?? '-' }}</td>
-              <td class="px-5 py-3.5 font-bold text-[#F37321]">{{ (o.items ?? []).reduce((s, i) => s + i.qty, 0).toLocaleString() }}</td>
-              <td class="px-5 py-3.5 font-semibold text-gray-700">{{ formatPrice((o.items ?? []).reduce((s, i) => s + i.unitPrice * i.qty, 0)) }}</td>
-              <td class="px-5 py-3.5">
-                <span class="text-xs font-bold px-2 py-0.5 rounded" :class="statusClass(o.status)">{{ o.status }}</span>
-              </td>
-            </tr>
-            <tr v-if="autoOrders.length === 0">
-              <td colspan="6" class="px-5 py-10 text-center text-sm text-gray-400">자동 발주 제안이 없습니다.</td>
-            </tr>
-          </tbody>
-        </table>
-      </div>
-    </div>
+    <!-- Tab Contents -->
+    <HqAutoOrderTable v-if="activeTab === 'auto'" :orders="autoOrders" @open-detail="openDetail" />
+    <HqManualOrderTable v-if="activeTab === 'manual'" :orders="pendingManualOrders" @open-detail="openDetail" />
+    <HqOrderHistoryTable v-if="activeTab === 'history'" :orders="orderHistory" @open-detail="openDetail" />
+    <HqAbnormalOrderTable v-if="activeTab === 'abnormal'" :orders="abnormalOrders"
+      @open-detail="openDetail" @approve="approveAbnormal" @reject="rejectAbnormal" />
 
-    <!-- Tab: 수동 발주 -->
-    <div v-if="activeTab === 'manual'" class="space-y-5">
-
-      <!-- ORDER_004: 진행 중인 수동 발주 목록 -->
-      <div>
-        <h3 class="text-sm font-bold text-gray-700 mb-2">진행 중인 수동 발주</h3>
-        <div class="bg-white border border-gray-200 rounded-lg overflow-hidden">
-          <table class="w-full text-sm text-left">
-            <thead>
-              <tr class="border-b border-gray-200 bg-gray-50">
-                <th class="px-5 py-3 text-[10px] font-bold text-gray-400 uppercase tracking-wider">발주번호</th>
-                <th class="px-5 py-3 text-[10px] font-bold text-gray-400 uppercase tracking-wider">대상 매장</th>
-                <th class="px-5 py-3 text-[10px] font-bold text-gray-400 uppercase tracking-wider">발주일시</th>
-                <th class="px-5 py-3 text-[10px] font-bold text-gray-400 uppercase tracking-wider">총 금액</th>
-                <th class="px-5 py-3 text-[10px] font-bold text-gray-400 uppercase tracking-wider">상태</th>
-              </tr>
-            </thead>
-            <tbody class="divide-y divide-gray-100">
-              <tr v-for="o in pendingManualOrders" :key="o.id" class="hover:bg-gray-50/50 transition-colors cursor-pointer" @click="openDetail({ id: o.id, type: '수동', store: o.store, status: o.status, date: o.date, items: o.items })">
-                <td class="px-5 py-3.5 font-mono text-xs text-gray-400">{{ o.id }}</td>
-                <td class="px-5 py-3.5 font-semibold text-gray-900">{{ o.store }}</td>
-                <td class="px-5 py-3.5 text-xs text-gray-400 font-mono">{{ o.date }}</td>
-                <td class="px-5 py-3.5 font-semibold text-gray-700">{{ formatPrice((o.items ?? []).reduce((s, i) => s + i.unitPrice * i.qty, 0)) }}</td>
-                <td class="px-5 py-3.5">
-                  <span class="text-xs font-bold px-2 py-0.5 rounded" :class="statusClass(o.status)">{{ o.status }}</span>
-                </td>
-              </tr>
-              <tr v-if="pendingManualOrders.length === 0">
-                <td colspan="5" class="px-5 py-8 text-center text-sm text-gray-400">진행 중인 수동 발주가 없습니다.</td>
-              </tr>
-            </tbody>
-          </table>
-        </div>
-      </div>
-
-    </div>
-
-    <!-- Tab: 이상 발주 (ORDER_007) -->
-    <div v-if="activeTab === 'abnormal'" class="space-y-4">
-      <div class="bg-white border border-gray-200 rounded-lg overflow-hidden">
-        <table class="w-full text-sm text-left">
-          <thead>
-            <tr class="border-b border-gray-200 bg-gray-50">
-              <th class="px-5 py-3 text-[10px] font-bold text-gray-400 uppercase tracking-wider">발주번호</th>
-              <th class="px-5 py-3 text-[10px] font-bold text-gray-400 uppercase tracking-wider">입점 매장</th>
-              <th class="px-5 py-3 text-[10px] font-bold text-gray-400 uppercase tracking-wider">발주수량</th>
-              <th class="px-5 py-3 text-[10px] font-bold text-gray-400 uppercase tracking-wider">평균수량</th>
-              <th class="px-5 py-3 text-[10px] font-bold text-gray-400 uppercase tracking-wider">초과율</th>
-              <th class="px-5 py-3 text-[10px] font-bold text-gray-400 uppercase tracking-wider">총 금액</th>
-              <th class="px-5 py-3 text-[10px] font-bold text-gray-400 uppercase tracking-wider">발주일시</th>
-              <th class="px-5 py-3 text-[10px] font-bold text-gray-400 uppercase tracking-wider">상태</th>
-              <th class="px-5 py-3 text-[10px] font-bold text-gray-400 uppercase tracking-wider text-center">처리</th>
-            </tr>
-          </thead>
-          <tbody class="divide-y divide-gray-100">
-            <tr v-for="o in abnormalOrders" :key="o.id"
-              class="hover:bg-gray-50/50 transition-colors cursor-pointer"
-              :class="o.processed ? 'opacity-50' : ''"
-              @click="openDetail({ id: o.id, type: '이상', store: o.store, status: o.processed ? '처리완료' : 'DANGER', date: o.date, items: o.items, avgQty: o.avgQty, ratio: o.ratio })"
-            >
-              <td class="px-5 py-3.5 font-mono text-xs text-gray-400">{{ o.id }}</td>
-              <td class="px-5 py-3.5 font-semibold text-gray-900">{{ o.store }}</td>
-              <td class="px-5 py-3.5 font-bold text-red-600">{{ o.qty.toLocaleString() }}</td>
-              <td class="px-5 py-3.5 text-gray-500">{{ o.avgQty.toLocaleString() }}</td>
-              <td class="px-5 py-3.5">
-                <span class="text-xs font-black px-2 py-0.5 rounded bg-red-50 text-red-600 border border-red-200">
-                  +{{ o.ratio }}%
-                </span>
-              </td>
-              <td class="px-5 py-3.5 font-semibold text-gray-700">{{ formatPrice((o.items ?? []).reduce((s, i) => s + i.unitPrice * i.qty, 0)) }}</td>
-              <td class="px-5 py-3.5 text-xs text-gray-400 font-mono">{{ o.date }}</td>
-              <td class="px-5 py-3.5">
-                <span class="text-xs font-bold px-2 py-0.5 rounded"
-                  :class="o.processed
-                    ? 'bg-gray-100 text-gray-400 border border-gray-200'
-                    : 'bg-red-50 text-red-600 border border-red-200'">
-                  {{ o.processed ? '처리완료' : 'DANGER' }}
-                </span>
-              </td>
-              <td class="px-5 py-3.5">
-                <div v-if="!o.processed" class="flex justify-center gap-1.5">
-                  <button @click.stop="approveAbnormal(o)"
-                    class="px-2.5 py-1 bg-[#F37321] text-white text-xs font-semibold hover:bg-[#e0661d] rounded cursor-pointer">승인</button>
-                  <button @click.stop="rejectAbnormal(o)"
-                    class="px-2.5 py-1 border border-gray-200 text-gray-600 text-xs font-medium hover:bg-gray-50 rounded cursor-pointer">반려</button>
-                </div>
-                <span v-else class="text-xs text-gray-400 block text-center">—</span>
-              </td>
-            </tr>
-          </tbody>
-        </table>
-      </div>
-    </div>
-
-    <!-- Tab: 발주 이력 (ORDER_006) -->
-    <div v-if="activeTab === 'history'" class="space-y-3">
-      <div class="bg-white border border-gray-200 rounded-lg px-5 py-4 flex flex-wrap gap-5 items-end">
-        <label class="flex flex-col gap-2">
-          <span class="text-[10px] font-bold text-gray-400 uppercase tracking-wider">유형</span>
-          <select v-model="historyFilterType"
-            class="w-24 px-3 py-2 rounded border border-gray-200 text-sm outline-none focus:border-[#F37321]">
-            <option value="">전체</option>
-            <option value="자동">자동</option>
-            <option value="수동">수동</option>
-          </select>
-        </label>
-        <label class="flex flex-col gap-2">
-          <span class="text-[10px] font-bold text-gray-400 uppercase tracking-wider">기간 시작</span>
-          <input v-model="historyDateFrom" type="date" class="px-3 py-2 rounded border border-gray-200 text-sm outline-none focus:border-[#F37321]" />
-        </label>
-        <label class="flex flex-col gap-2">
-          <span class="text-[10px] font-bold text-gray-400 uppercase tracking-wider">기간 종료</span>
-          <input v-model="historyDateTo" type="date" class="px-3 py-2 rounded border border-gray-200 text-sm outline-none focus:border-[#F37321]" />
-        </label>
-        <div class="flex flex-col gap-2 flex-1 min-w-40">
-          <span class="text-[10px] font-bold text-gray-400 uppercase tracking-wider">검색</span>
-          <div class="flex gap-2 items-center">
-            <input v-model="historySearch" type="search" placeholder="발주번호·가맹점·품목"
-              class="flex-1 px-3 py-2 rounded border border-gray-200 text-sm outline-none focus:border-[#F37321]" />
-            <button type="button" class="text-xs font-semibold text-gray-500 border border-gray-200 px-4 py-2 rounded hover:bg-gray-50 shrink-0 cursor-pointer"
-              @click="historyFilterType = ''; historyDateFrom = ''; historyDateTo = ''; historySearch = ''">
-              초기화
-            </button>
-          </div>
-        </div>
-      </div>
-      <div class="bg-white border border-gray-200 rounded-lg overflow-hidden">
-        <table class="w-full text-sm text-left">
-          <thead>
-            <tr class="border-b border-gray-200 bg-gray-50">
-              <th class="px-5 py-3 text-[10px] font-bold text-gray-400 uppercase tracking-wider">발주번호</th>
-              <th class="px-5 py-3 text-[10px] font-bold text-gray-400 uppercase tracking-wider">유형</th>
-              <th class="px-5 py-3 text-[10px] font-bold text-gray-400 uppercase tracking-wider">입점 매장</th>
-              <th class="px-5 py-3 text-[10px] font-bold text-gray-400 uppercase tracking-wider">발주일시</th>
-              <th class="px-5 py-3 text-[10px] font-bold text-gray-400 uppercase tracking-wider">총 금액</th>
-              <th class="px-5 py-3 text-[10px] font-bold text-gray-400 uppercase tracking-wider">상태</th>
-            </tr>
-          </thead>
-          <tbody class="divide-y divide-gray-100">
-            <tr v-for="h in filteredOrderHistory" :key="h.id" class="hover:bg-gray-50/50 transition-colors cursor-pointer" @click="openDetail({ id: h.id, type: h.type, store: h.store, status: h.status, date: h.date, items: h.items })">
-              <td class="px-5 py-3.5 font-mono text-xs text-gray-400">{{ h.id }}</td>
-              <td class="px-5 py-3.5">
-                <span class="text-xs font-bold px-2 py-0.5 rounded"
-                  :class="h.type === '자동' ? 'bg-blue-50 text-blue-600 border border-blue-200' : 'bg-purple-50 text-purple-600 border border-purple-200'">
-                  {{ h.type }}
-                </span>
-              </td>
-              <td class="px-5 py-3.5 font-semibold text-gray-900">{{ h.store }}</td>
-              <td class="px-5 py-3.5 text-xs text-gray-400 font-mono">{{ h.date }}</td>
-              <td class="px-5 py-3.5 font-semibold text-gray-700">{{ formatPrice((h.items ?? []).reduce((s, i) => s + i.unitPrice * i.qty, 0)) }}</td>
-              <td class="px-5 py-3.5">
-                <span class="text-xs font-bold px-2 py-0.5 rounded"
-                  :class="statusClass(h.status)">{{ h.status }}</span>
-              </td>
-            </tr>
-            <tr v-if="filteredOrderHistory.length === 0">
-              <td colspan="6" class="px-5 py-10 text-center text-sm text-gray-400">조건에 맞는 발주 이력이 없습니다.</td>
-            </tr>
-          </tbody>
-        </table>
-      </div>
-    </div>
-
-    <!-- 수동 발주 생성 Modal (ORDER_003) -->
-    <div v-if="showManualForm" class="fixed inset-0 z-50 flex items-center justify-center">
-      <div class="absolute inset-0 bg-black/40" @click="showManualForm = false"></div>
-      <div class="relative bg-white rounded-xl w-full max-w-lg border border-gray-200 shadow-xl">
-        <div class="px-6 py-4 border-b border-gray-200 flex justify-between items-center">
-          <h3 class="font-bold text-gray-900">수동 발주 생성</h3>
-          <button @click="showManualForm = false" class="text-gray-400 hover:text-gray-600 cursor-pointer">✕</button>
-        </div>
-        <div class="p-6 space-y-4">
-          <div class="space-y-1.5">
-            <label class="text-xs font-bold text-gray-500 uppercase tracking-wider">대상 매장</label>
-            <select v-model="manualForm.store"
-              class="w-full px-3 py-2 rounded border border-gray-200 text-sm focus:border-[#F37321] focus:ring-2 focus:ring-[#F37321]/10 outline-none">
-              <option value="">선택</option>
-              <option>한우 오마카세</option>
-              <option>이탈리안 키친</option>
-              <option>일식 스시바</option>
-              <option>차이나 가든</option>
-            </select>
-          </div>
-          <div class="space-y-2">
-            <div class="flex justify-between items-center">
-              <label class="text-xs font-bold text-gray-500 uppercase tracking-wider">발주 품목</label>
-              <button @click="addManualItem" class="text-xs text-[#F37321] font-semibold hover:underline flex items-center gap-1 cursor-pointer">
-                <Plus class="w-3 h-3" /> 품목 추가
-              </button>
-            </div>
-            <div v-for="(item, idx) in manualForm.items" :key="idx" class="flex gap-2 items-center">
-              <select v-model="item.product" @change="item.unitPrice = productPrices[item.product] ?? 0"
-                class="flex-1 px-3 py-2 rounded border border-gray-200 text-sm focus:border-[#F37321] focus:ring-2 focus:ring-[#F37321]/10 outline-none">
-                <option value="">품목 선택</option>
-                <option>한우 등심</option>
-                <option>올리브오일</option>
-                <option>버터</option>
-                <option>생크림</option>
-                <option>간장</option>
-                <option>양파</option>
-                <option>생수</option>
-              </select>
-              <div class="flex items-center gap-1.5">
-                <input v-model.number="item.qty" type="number" min="1" placeholder="수량"
-                  class="w-20 px-3 py-2 rounded border border-gray-200 text-sm focus:border-[#F37321] focus:ring-2 focus:ring-[#F37321]/10 outline-none" />
-                <span class="text-xs text-gray-400 font-medium w-6">{{ PRODUCT_UNIT[item.product] ?? '' }}</span>
-              </div>
-              <div class="w-28 px-3 py-2 rounded border border-gray-100 bg-gray-50 text-sm text-gray-500 text-right shrink-0">
-                {{ item.product ? formatPrice(productPrices[item.product] ?? 0) : '-' }}
-              </div>
-              <button @click="manualForm.items.splice(idx, 1)"
-                class="px-3 py-2 text-xs font-semibold rounded border border-red-200 text-red-500 bg-red-50 hover:bg-red-500 hover:text-white hover:cursor-pointer transition-colors shrink-0">
-                삭제
-              </button>
-            </div>
-            <div v-if="manualForm.items.length === 0"
-              class="text-sm text-gray-400 text-center py-4 bg-gray-50 border border-gray-100 rounded">
-              품목 추가 버튼을 눌러 발주 품목을 입력하세요.
-            </div>
-            <div v-if="manualForm.items.length > 0" class="text-right text-sm font-bold text-[#F37321]">
-              합계: {{ formatPrice(manualForm.items.reduce((s, i) => s + (i.unitPrice || 0) * (i.qty || 0), 0)) }}
-            </div>
-          </div>
-        </div>
-        <div class="px-6 py-4 border-t border-gray-100 flex gap-3">
-          <button @click="showManualForm = false"
-            class="flex-1 py-2.5 rounded border border-gray-200 text-sm font-semibold text-gray-600 hover:bg-gray-50 cursor-pointer">취소</button>
-          <button @click="submitManualOrder"
-            class="flex-1 py-2.5 rounded bg-[#F37321] text-white text-sm font-bold hover:bg-[#e0661d] cursor-pointer">발주 생성</button>
-        </div>
-      </div>
-    </div>
-
-    <!-- 발주 상세 조회 Modal (ORDER_005) -->
-    <div v-if="showDetail" class="fixed inset-0 z-50 flex items-center justify-center">
-      <div class="absolute inset-0 bg-black/40" @click="showDetail = false"></div>
-      <div class="relative bg-white rounded-xl w-full max-w-lg border border-gray-200 shadow-xl">
-        <div class="px-6 py-4 border-b border-gray-200 flex justify-between items-center">
-          <div>
-            <h3 class="font-bold text-gray-900">발주 상세</h3>
-            <p class="text-xs text-gray-400 font-mono mt-0.5">{{ selectedOrder?.id }}</p>
-          </div>
-          <button @click="showDetail = false" class="text-gray-400 hover:text-gray-600 cursor-pointer">✕</button>
-        </div>
-        <div v-if="selectedOrder" class="p-6 space-y-4">
-          <div class="grid grid-cols-2 gap-4 text-sm">
-            <div>
-              <p class="text-xs text-gray-400 mb-1">유형</p>
-              <span class="text-xs font-bold px-2 py-0.5 rounded border"
-                :class="selectedOrder.type === '자동' ? 'bg-blue-50 text-blue-700 border-blue-200' : 'bg-purple-50 text-purple-700 border-purple-200'">
-                {{ selectedOrder.type }}
-              </span>
-            </div>
-            <div>
-              <p class="text-xs text-gray-400 mb-1">상태</p>
-              <span class="text-xs font-bold px-2 py-0.5 rounded" :class="statusClass(selectedOrder.status)">
-                {{ selectedOrder.status }}
-              </span>
-            </div>
-            <div>
-              <p class="text-xs text-gray-400 mb-1">대상 매장</p>
-              <p class="font-semibold text-gray-900">{{ selectedOrder.store }}</p>
-            </div>
-            <div>
-              <p class="text-xs text-gray-400 mb-1">발주 일시</p>
-              <p class="text-gray-700 font-mono text-xs">{{ selectedOrder.date }}</p>
-            </div>
-          </div>
-          <div>
-            <p class="text-xs text-gray-400 mb-2">품목 목록</p>
-            <div class="border border-gray-100 rounded-lg overflow-hidden">
-              <table class="w-full text-sm">
-                <thead class="bg-gray-50">
-                  <tr>
-                    <th class="px-4 py-2.5 text-left text-[10px] font-bold text-gray-400 uppercase">품목명</th>
-                    <th class="px-4 py-2.5 text-right text-[10px] font-bold text-gray-400 uppercase">수량</th>
-                    <template v-if="selectedOrder.type === '이상'">
-                      <th class="px-4 py-2.5 text-right text-[10px] font-bold text-gray-400 uppercase">평균수량</th>
-                      <th class="px-4 py-2.5 text-right text-[10px] font-bold text-gray-400 uppercase">초과율</th>
-                    </template>
-                    <template v-else>
-                      <th class="px-4 py-2.5 text-right text-[10px] font-bold text-gray-400 uppercase">단가</th>
-                      <th class="px-4 py-2.5 text-right text-[10px] font-bold text-gray-400 uppercase">금액</th>
-                    </template>
-                  </tr>
-                </thead>
-                <tbody class="divide-y divide-gray-100">
-                  <tr v-for="item in selectedOrder.items" :key="item.product"
-                    :class="item.isAbnormal ? 'bg-red-50/60' : ''">
-                    <td class="px-4 py-2.5">
-                      <div class="flex items-center gap-1.5">
-                        <span :class="item.isAbnormal ? 'text-red-700 font-semibold' : 'text-gray-800'">{{ item.product }}</span>
-                        <span v-if="item.isAbnormal" class="text-[10px] font-black px-1.5 py-0.5 rounded bg-red-100 text-red-600 border border-red-200">이상</span>
-                      </div>
-                    </td>
-                    <td class="px-4 py-2.5 text-right font-semibold" :class="item.isAbnormal ? 'text-red-600' : 'text-gray-900'">{{ item.qty.toLocaleString() }}</td>
-                    <template v-if="selectedOrder.type === '이상'">
-                      <td class="px-4 py-2.5 text-right text-gray-500">{{ item.avgQty?.toLocaleString() ?? '-' }}</td>
-                      <td class="px-4 py-2.5 text-right">
-                        <span v-if="item.isAbnormal" class="text-xs font-black px-2 py-0.5 rounded bg-red-50 text-red-600 border border-red-200">+{{ item.ratio }}%</span>
-                        <span v-else class="text-xs text-gray-400">+{{ item.ratio }}%</span>
-                      </td>
-                    </template>
-                    <template v-else>
-                      <td class="px-4 py-2.5 text-right text-xs text-gray-500">{{ item.unitPrice ? item.unitPrice.toLocaleString() + '원' : '-' }}</td>
-                      <td class="px-4 py-2.5 text-right font-bold text-[#F37321]">{{ item.unitPrice ? formatPrice(item.unitPrice * item.qty) : '-' }}</td>
-                    </template>
-                  </tr>
-                </tbody>
-                <tfoot v-if="selectedOrder.type !== '이상'" class="bg-gray-50 border-t border-gray-200">
-                  <tr>
-                    <td colspan="3" class="px-4 py-2.5 text-right text-xs font-bold text-gray-500">합계</td>
-                    <td class="px-4 py-2.5 text-right font-black text-[#F37321]">
-                      {{ formatPrice((selectedOrder.items ?? []).reduce((s, i) => s + (i.unitPrice ? i.unitPrice * i.qty : 0), 0)) }}
-                    </td>
-                  </tr>
-                </tfoot>
-              </table>
-            </div>
-          </div>
-          <div v-if="selectedOrder.avgQty" class="grid grid-cols-2 gap-4 text-sm p-3 bg-rose-50 rounded-lg border border-rose-100">
-            <div>
-              <p class="text-xs text-rose-400 mb-1">평균 발주수량</p>
-              <p class="font-semibold text-gray-800">{{ selectedOrder.avgQty.toLocaleString() }}</p>
-            </div>
-            <div>
-              <p class="text-xs text-rose-400 mb-1">초과율</p>
-              <p class="font-bold text-red-600">+{{ selectedOrder.ratio }}%</p>
-            </div>
-          </div>
-          <div v-if="selectedOrder.note" class="text-sm">
-            <p class="text-xs text-gray-400 mb-1">비고</p>
-            <p class="text-gray-600 bg-gray-50 px-3 py-2 rounded border border-gray-100">{{ selectedOrder.note }}</p>
-          </div>
-        </div>
-        <div class="px-6 py-4 border-t border-gray-100 flex justify-end">
-          <button @click="showDetail = false"
-            class="px-4 py-2 text-sm font-semibold text-gray-600 border border-gray-200 rounded hover:bg-gray-50 cursor-pointer">닫기</button>
-        </div>
-      </div>
-    </div>
-
-    <!-- 이상 발주 기준 설정 Modal (ORDER_008) -->
-    <div v-if="showSettings" class="fixed inset-0 z-50 flex items-center justify-center">
-      <div class="absolute inset-0 bg-black/40" @click="showSettings = false"></div>
-      <div class="relative bg-white rounded-lg w-full max-w-sm border border-gray-200 shadow-xl">
-        <div class="px-6 py-4 border-b border-gray-200 flex justify-between items-center">
-          <h3 class="font-bold text-gray-900">이상 발주 기준 설정</h3>
-          <button @click="showSettings = false" class="text-gray-400 hover:text-gray-600 cursor-pointer">✕</button>
-        </div>
-        <div class="p-6 space-y-5">
-          <div class="space-y-3">
-            <label class="text-xs font-bold text-gray-500 uppercase tracking-wider">이상 발주 감지 기준</label>
-            <div class="space-y-3">
-              <div>
-                <p class="text-xs text-gray-500 mb-1.5 font-medium">초과 비율 기준</p>
-                <div class="flex items-center gap-3">
-                  <input v-model.number="abnormalThreshold" type="number" min="100" max="1000" step="10"
-                    class="w-28 px-3 py-2 rounded border border-gray-200 text-sm font-bold focus:border-[#F37321] focus:ring-2 focus:ring-[#F37321]/10 outline-none" />
-                  <span class="text-sm text-gray-600">% 초과 시 이상 감지</span>
-                </div>
-              </div>
-              <div>
-                <p class="text-xs text-gray-500 mb-1.5 font-medium">분석 기준 기간</p>
-                <div class="flex items-center gap-3">
-                  <input v-model.number="abnormalMonths" type="number" min="1" max="24"
-                    class="w-28 px-3 py-2 rounded border border-gray-200 text-sm font-bold focus:border-[#F37321] focus:ring-2 focus:ring-[#F37321]/10 outline-none" />
-                  <span class="text-sm text-gray-600">개월 과거 데이터 기준</span>
-                </div>
-              </div>
-            </div>
-            <p class="text-xs text-gray-400 bg-gray-50 px-3 py-2 rounded border border-gray-100">
-              현재 설정: 최근 <strong class="text-gray-700">{{ abnormalMonths }}개월</strong> 평균 대비
-              <strong class="text-red-600">{{ abnormalThreshold }}%</strong> 이상 초과 발주 감지
-            </p>
-          </div>
-          <div class="flex gap-3 pt-1">
-            <button @click="showSettings = false"
-              class="flex-1 py-2.5 rounded border border-gray-200 text-sm font-semibold text-gray-600 hover:bg-gray-50 cursor-pointer">취소</button>
-            <button @click="showSettings = false"
-              class="flex-1 py-2.5 rounded bg-[#F37321] text-white text-sm font-bold hover:bg-[#e0661d] cursor-pointer">저장</button>
-          </div>
-        </div>
-      </div>
-    </div>
+    <!-- Modals -->
+    <HqManualOrderModal :visible="showManualForm" @close="showManualForm = false" @submit="submitManualOrder" />
+    <HqOrderDetailModal :order="selectedOrder" @close="selectedOrder = null" />
+    <HqDangerSettingsModal :visible="showSettings" @close="showSettings = false" @save="saveDangerSettings" />
   </div>
 </template>
 
 <script setup>
 import { ref, computed, onMounted } from 'vue'
 import { useRoute, useRouter } from 'vue-router'
-import { Plus, X, Settings } from 'lucide-vue-next'
+import { Plus, Settings } from 'lucide-vue-next'
+import { productPrices } from '@/components/orders/orderUtils'
+import HqAutoOrderTable from '@/components/orders/HqAutoOrderTable.vue'
+import HqManualOrderTable from '@/components/orders/HqManualOrderTable.vue'
+import HqOrderHistoryTable from '@/components/orders/HqOrderHistoryTable.vue'
+import HqAbnormalOrderTable from '@/components/orders/HqAbnormalOrderTable.vue'
+import HqOrderDetailModal from '@/components/orders/HqOrderDetailModal.vue'
+import HqManualOrderModal from '@/components/orders/HqManualOrderModal.vue'
+import HqDangerSettingsModal from '@/components/orders/HqDangerSettingsModal.vue'
 
 const route = useRoute()
 const router = useRouter()
-
-const PRODUCT_UNIT = {
-  '한우 등심':   'kg',
-  '올리브오일':  'L',
-  '버터':        'kg',
-  '생크림':      'L',
-  '간장':        'L',
-  '양파':        'kg',
-  '생수':        '박스',
-}
-
-const productPrices = {
-  '한우 등심':   85000,
-  '올리브오일':  12000,
-  '버터':        9000,
-  '생크림':      7000,
-  '간장':        4000,
-  '양파':        1500,
-  '생수':        8000,
-}
-
-function formatPrice(n) {
-  return '₩ ' + n.toLocaleString('ko-KR')
-}
 
 const abnormalCount = computed(() => abnormalOrders.value.filter(o => !o.processed).length)
 
@@ -501,7 +73,6 @@ const tabs = computed(() => [
   { id: 'abnormal', label: '이상 발주', badge: abnormalCount.value || null },
 ])
 const activeTab = ref('auto')
-
 
 const autoOrders = ref([
   { id: 'AUTO-20260413-001', store: '이탈리안 키친',  date: '2026-04-13 22:00', status: '제안중',
@@ -527,24 +98,6 @@ const orderHistory = ref([
     items: [{ product: '연어', qty: 10, unitPrice: 32000 }, { product: '마늘', qty: 5, unitPrice: 8000 }] },
 ])
 
-const historyFilterType = ref('')
-const historyDateFrom = ref('')
-const historyDateTo = ref('')
-const historySearch = ref('')
-
-const filteredOrderHistory = computed(() =>
-  orderHistory.value.filter((h) => {
-    if (historyFilterType.value && h.type !== historyFilterType.value) return false
-    const day = h.date.slice(0, 10)
-    if (historyDateFrom.value && day < historyDateFrom.value) return false
-    if (historyDateTo.value && day > historyDateTo.value) return false
-    const q = historySearch.value.trim()
-    if (q && !h.id.includes(q) && !h.store.includes(q) && !h.items.some(i => i.product.includes(q))) return false
-    return true
-  }),
-)
-
-// 이상 발주 데이터 (ORDER_007)
 const abnormalOrders = ref([
   { id: 'ORD-20260414-ABN1', store: '이탈리안 키친', qty: 85, avgQty: 15, ratio: 567, date: '2026-04-14 11:22', processed: false,
     items: [
@@ -563,23 +116,18 @@ const abnormalOrders = ref([
     ] },
 ])
 
+const pendingManualOrders = ref([
+  { id: 'MAN-20260420-001', store: '이탈리안 키친', date: '2026-04-20 09:10', status: '확정',
+    items: [{ product: '한우 등심', qty: 10, unitPrice: 85000 }, { product: '버터', qty: 5, unitPrice: 9000 }] },
+  { id: 'MAN-20260419-002', store: '일식 스시바',   date: '2026-04-19 14:30', status: '배송중',
+    items: [{ product: '올리브오일', qty: 8, unitPrice: 12000 }, { product: '간장', qty: 5, unitPrice: 4000 }, { product: '생수', qty: 10, unitPrice: 8000 }] },
+])
 
+// Tab routing
 function applyOrderRouteQuery() {
   const q = route.query
-  if (q.tab === 'abnormal') {
-    activeTab.value = 'abnormal'
-    return
-  }
-  if (q.tab === 'manual') {
-    activeTab.value = 'manual'
-    return
-  }
-  if (q.tab === 'history') {
-    activeTab.value = 'history'
-    return
-  }
-  if (q.tab === 'auto') {
-    activeTab.value = 'auto'
+  if (q.tab && ['auto', 'manual', 'history', 'abnormal'].includes(q.tab)) {
+    activeTab.value = q.tab
   }
 }
 
@@ -589,88 +137,42 @@ onMounted(() => {
 
 function setOrderViewTab(id) {
   activeTab.value = id
-  if (id === 'auto') {
-    router.replace({ path: '/order', query: { tab: 'auto' } })
-    return
-  }
-  if (id === 'abnormal') {
-    router.replace({ path: '/order', query: { tab: 'abnormal' } })
-    return
-  }
-  if (id === 'manual') {
-    router.replace({ path: '/order', query: { tab: 'manual' } })
-    return
-  }
-  if (id === 'history') {
-    router.replace({ path: '/order', query: { tab: 'history' } })
-  }
+  router.replace({ path: '/order', query: { tab: id } })
 }
 
-// 발주 상세 조회 (ORDER_005)
-const showDetail = ref(false)
+// Detail modal
 const selectedOrder = ref(null)
 
 function openDetail(order) {
   selectedOrder.value = order
-  showDetail.value = true
 }
 
-// 이상 발주 기준 설정 (ORDER_008)
+// Settings modal
 const showSettings = ref(false)
-const abnormalThreshold = ref(200) // 평균 대비 %
-const abnormalMonths = ref(3)      // 분석 기준 개월 수
 
-// 이상 발주 처리
+function saveDangerSettings({ threshold, months }) {
+  showSettings.value = false
+}
+
+// Abnormal order actions
 function approveAbnormal(o) { o.processed = true; alert(`${o.store} 발주 승인 처리되었습니다.`) }
 function rejectAbnormal(o)  { o.processed = true; alert(`${o.store} 발주 반려 처리되었습니다.`) }
 
-const pendingManualOrders = ref([
-  { id: 'MAN-20260420-001', store: '이탈리안 키친', date: '2026-04-20 09:10', status: '확정',
-    items: [{ product: '한우 등심', qty: 10, unitPrice: 85000 }, { product: '버터', qty: 5, unitPrice: 9000 }] },
-  { id: 'MAN-20260419-002', store: '일식 스시바',   date: '2026-04-19 14:30', status: '배송중',
-    items: [{ product: '올리브오일', qty: 8, unitPrice: 12000 }, { product: '간장', qty: 5, unitPrice: 4000 }, { product: '생수', qty: 10, unitPrice: 8000 }] },
-])
-
+// Manual order modal
 const showManualForm = ref(false)
-const manualForm = ref({ store: '', items: [] })
 
-function addManualItem() {
-  manualForm.value.items.push({ product: '', qty: 1, unitPrice: 0 })
-}
-
-function submitManualOrder() {
-  if (!manualForm.value.store || manualForm.value.items.length === 0) {
-    alert('매장과 품목을 입력해주세요.')
-    return
-  }
+function submitManualOrder({ store, items }) {
   const now = new Date().toISOString().slice(0, 16).replace('T', ' ')
-  const validItems = manualForm.value.items.filter(i => i.product)
-  if (validItems.length > 0) {
-    orderHistory.value.unshift({
-      id: `ORD-${Date.now()}`,
-      type: '수동',
-      store: manualForm.value.store,
-      date: now,
-      status: '확정',
-      items: validItems.map(i => ({ product: i.product, qty: i.qty, unitPrice: productPrices[i.product] ?? 0 })),
-    })
-  }
+  orderHistory.value.unshift({
+    id: `ORD-${Date.now()}`,
+    type: '수동',
+    store,
+    date: now,
+    status: '확정',
+    items,
+  })
   alert('발주가 생성되었습니다.')
-  manualForm.value = { store: '', items: [] }
   showManualForm.value = false
   setOrderViewTab('history')
-}
-
-
-function statusClass(status) {
-  const map = {
-    '제안중':   'bg-blue-50 text-blue-600 border border-blue-200',
-    '검토필요': 'bg-rose-50 text-rose-700 border border-rose-200',
-    '확정':     'bg-green-50 text-green-700 border border-green-200',
-    '거절':     'bg-gray-100 text-gray-400 border border-gray-200',
-    '배송중':   'bg-orange-50 text-orange-600 border border-orange-200',
-    '입고완료': 'bg-emerald-50 text-emerald-700 border border-emerald-200',
-  }
-  return map[status] || 'bg-gray-100 text-gray-500 border border-gray-200'
 }
 </script>
